Remove lifetime from EventPump, move into `sdl2::sdl` module
The lifetime on EventPump isn't really needed anymore.
